Change int in model and other occurrence to double

+++ Freestyle.UndercoverDev/Data/Models/MatchData.cs
@@ -6,15 +6,15 @@ public class MatchData
     public string? League { get; set;} = "";
     public string? HomeTeam { get; set; } = "";
     public string? AwayTeam { get; set; } = "";
-    public float? HomeWin { get; set; }
-    public float? Draw { get; set; }
-    public float? AwayWin { get; set; }
-    public float? OverOneGoal { get; set; }
-    public float? OverTwoGoals { get; set; }
-    public float? OverThreeGoals { get; set; }
-    public float? OverFourGoals { get; set; }
-    public float? UnderOneGoal { get; set; }
-    public float? UnderTwoGoals { get; set; }
-    public float? UnderThreeGoals { get; set; }
-    public float? UnderFourGoals { get; set; }
+    public double? HomeWin { get; set; }
+    public double? Draw { get; set; }
+    public double? AwayWin { get; set; }
+    public double? OverOneGoal { get; set; }
+    public double? OverTwoGoals { get; set; }
+    public double? OverThreeGoals { get; set; }
+    public double? OverFourGoals { get; set; }
+    public double? UnderOneGoal { get; set; }
+    public double? UnderTwoGoals { get; set; }
+    public double? UnderThreeGoals { get; set; }
+    public double? UnderFourGoals { get; set; }
 }
